first end-to-end (brittle) coordination
VendingService is now an HTTP REST/Json service that orchestrates the call between the bank and the stock
Basic error handling/forwarding is in place, with some tentative compensation in case of error.

Calls are idempotent, but never retried (yet).

Also, this could leave the state inconsistent in case the compensation mechanism itself fails (which is unlikely since it consists in increasing the sock, though anything is possible)